curly brace standard
applied standard for curly braces to be on their own line
