+++ Lab_03/Lab_03_D2.c
@@ -29,7 +29,8 @@ void main(void)
     P1OUT &= ~BIT0;             // LED2 is OFF
     P4OUT |= BIT7;              // LED1 is ON
     unsigned int i = 0;
-    while(1) {                       // Infinite loop
+    while(1)
+	{                       // Infinite loop
         for (i = 0; i < 50000; i++); // Delay 0.5s
                                      // 0.5s on, 0.5s off => 1/(1s) = 1Hz
         P1OUT ^= BIT0;               // Toggle LED1
