+++ Lab_06/Lab_06_D5.c
@@ -22,7 +22,8 @@
 
 unsigned char S2pressed = 0;        // S2 status (0 not pressed, 1 pressed)
 
-void main(void) {
+void main(void)
+{
     WDTCTL = WDTPW+WDTHOLD;         // Stop WDT
     P1DIR |= BIT0;                  // Set LED1 as output
     P1OUT = 0x00;                   // Clear LED1 status
@@ -43,13 +44,16 @@ void main(void) {
 
 // Port 2 interrupt service routine
 #pragma vector = PORT1_VECTOR
-__interrupt void Port1_ISR (void) {
+__interrupt void Port1_ISR (void)
+{
     if (S2pressed == 0) {
         S2pressed = 1;
         P1OUT |= BIT0;                // LED1 is turned ON
         P1IFG &= ~BIT1;               // P1IFG.BIT0 is cleared
         P1IES &= ~BIT1;               // P1IES.BIT0 low/high edge
-    } else if (S2pressed == 1) {
+    }
+	else if (S2pressed == 1)
+	{
         S2pressed = 0;
         P1OUT &= ~BIT0;                // LED1 is turned ON
         P1IFG &= ~BIT1;                // P1IFG.BIT0 is cleared
