+++ Lab_06/Lab_06_D8.c
@@ -100,7 +100,8 @@ __interrupt void PORT2_ISR(void)
         // if not at 8Mhz, let us change to 8Mhz
         change_clock_freq_8Mhz();
         is8Mhz = 1;
-    }else
+    }
+	else
     {
         // if already in 8Mhz, let us take back to 1Mhz
         change_clock_freq_1Mhz();
