+++ Lab_07/Lab_07_D5.c
@@ -33,7 +33,8 @@
  *------------------------------------------------------------------------------*/
 #include <msp430F5529.h>
 
-void main(void) {
+void main(void)
+{
     WDTCTL = WDTPW +WDTHOLD;    // Stop WDT
 
     P2DIR |= BIT4;              // P2.4 output
